moved git functions to corresponding file
